update package manager text
